feat: sidebar progress stats added and refactor
